api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Michael Caley <mjcaley@darkarctic.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
azure-cosmosdb-table = "^1.0.6"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
api = "api.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
